# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Survey Service OpenApi3 Documentation",
        "description": "Behandam Survey Service Api documentation",
        "contact": {
            "email": "admin@kermany.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "LocalHost API Server"
        },
        {
            "url": "https://debug.behaminplus.ir/survey-service",
            "description": "Debug API Server"
        },
        {
            "url": "https://develop.behaminplus.ir/survey-service",
            "description": "DEVELOP API Server"
        }
    ],
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "get list of all categories",
                "operationId": "getListOfCategories",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "define a new category",
                "operationId": "postANewCategories",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/v1/categories/{categoryId}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "get category item by id",
                "operationId": "getCategoryItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID of category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "update category item",
                "operationId": "updateCategoryItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID of category item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "remove item by id",
                "operationId": "removeItemById",
                "parameters": [
                    {
                        "name": "X-Proxy-Token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "4fVB9SZidiBAADD2444nLZxxbWk92UcPQkwM8k"
                    },
                    {
                        "name": "categoryId",
                        "in": "path",
                        "description": "ID of category item",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoryStoreRequest": {
                "title": "CategoryStoreRequest",
                "description": "Category Store Request body data",
                "required": [
                    "title"
                ],
                "type": "object",
                "example": {
                    "title": "YOUR TITLE OF ..."
                }
            },
...
        },
        "securitySchemes": {
            "access-token": {
                "type": "http",
                "description": "Use Custom access token",
                "name": "AccessToken",
                "scheme": "access-token"
            }
        }
    },
...
    "security": [
        []
    ]
}